+++ lib/src/tafsir/tafsir.dart
@@ -3,13 +3,9 @@ import 'dart:developer' show log;
 import 'dart:io' show File, Directory;
 
 import 'package:dio/dio.dart';
-import 'package:drift/drift.dart' as drift;
-import 'package:drift/native.dart' show NativeDatabase;
-// ignore: implementation_imports
-import 'package:drift/src/runtime/api/runtime_api.dart';
 import 'package:flutter/foundation.dart' show kDebugMode;
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart' show rootBundle, ByteData;
+import 'package:flutter/services.dart' show rootBundle;
 import 'package:get/get.dart' hide Response;
 import 'package:get_storage/get_storage.dart';
 import 'package:path/path.dart';